+++ src/wallet/wallet2_basic/README.md
@@ -1,0 +1,3 @@
+# `wallet2_basic`
+
+The `wallet2_basic` library is a minimalistic library for loading / storing historic `wallet2` files.
