+++ tests/fuzz/cold-transaction.cpp
@@ -32,6 +32,7 @@
 #include "cryptonote_basic/cryptonote_basic.h"
 #include "cryptonote_basic/cryptonote_format_utils.h"
 #include "wallet/wallet2.h"
+#include "wallet/wallet2_basic/wallet2_serialization.h"
 #include "fuzzer.h"
 
 static tools::wallet2 *wallet = NULL;
